Update to use List instead of Array and add new mapping functions
This commit updates the code to use Lists instead of Arrays for methods that accept collections. It also includes new mapping functions to convert internal representations into structures usable by JavaScript and WebAssembly portions of the application. Furthermore, buffer size conversion to JavaScript numbers are added. These changes aim to improve resource handling and make the code easier to maintain and understand.

+++ wgpu4k/src/wasmJsMain/kotlin/io.ygdrasil.wgpu/Buffer.wasmJs.kt
@@ -5,7 +5,7 @@ import io.ygdrasil.wgpu.internal.js.GPUBuffer
 actual class Buffer(internal val handler: GPUBuffer) : AutoCloseable {
 
     actual val size: GPUSize64
-        get() = handler.size
+        get() = handler.size.toLong()
     actual val usage: Set<BufferUsage>
         get() = BufferUsage.entries.filter { it.value and handler.usage != 0 }.toSet()
     actual val mapState: BufferMapState
